dream_merch_maker/requirement.txt
Pillow==8.1.2
